changes functions to be more kotlin friendly, added helper functions for item creation

+++ src/main/kotlin/me/clip/voteparty/conf/ConfigVoteParty.kt
@@ -65,7 +65,7 @@ data class ConfigVoteParty(var settings: SettingsConfig?, var effects: EffectsCo
 	) : Config
 	
 	data class CommandsReward(
-			var delay: Int,
+			var delay: Long,
 			var commands: List<Command>
 	) : Config
